Refactor: user's page 🪄
