+++ src/pages/User/CreateUserPage.tsx
@@ -4,7 +4,7 @@ import { useCallback, useContext, useMemo } from "react"
 import { I18nContext } from "../../contexts/i18n.context"
 import { CommonLabels } from "../../i18n/commonLabels.i18n"
 import { useNavigate } from "react-router-dom"
-import { UserForm } from "./form/UserForm"
+import { UserForm } from "./components"
 import { useUser } from "./hooks/useUser"
 import { createUserPayload } from "../../types/user.type"
 import { useAlert } from "../../hooks/useAlert"
