+++ src/pages/User/components/UserDeleteModal.type.d.ts
@@ -1,0 +1,7 @@
+export interface IUserDeleteModalProps {
+    isOpen: boolean;
+    onClose: () => void;
+    onConfirm: (id: number) => void;
+    userId: number | null;
+    isLoading: boolean;
+}
