+++ src/pages/User/components/UserTable.type.d.ts
@@ -1,0 +1,8 @@
+import { User } from "../../../types/user.type"
+
+export interface IUserTableProps {
+    users: User[]
+    isLoading: boolean
+    onUserEditClick: (id: number) => void
+    onUserDeleteClick: (id: number) => void
+}
