+++ src/pages/User/components/index.ts
@@ -1,0 +1,3 @@
+export * from './UserDeleteModal'
+export * from './UserForm'
+export * from './UserTable'
