+++ src/pages/User/hooks/useUserList.type.d.ts
@@ -1,0 +1,7 @@
+import { findAllUsersResponse } from "../../../types/user.type"
+
+export interface IUseUserList {
+    users: findAllUsersResponse
+    isLoading: boolean
+    refreshRequest: () => void
+}
